backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
aiosqlite>=0.19.0

# Validation
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6

# HTTP client
httpx>=0.27.0

# Mistral AI
mistralai>=1.0.0

# LangChain (Phase 5+, install now to avoid rebuild)
langchain>=0.3.0
langchain-mistralai>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0

# MCP client (Live Slack via @modelcontextprotocol/server-slack)
mcp>=1.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
